cleanups, add sequence with multiplicity

+++ metagraph/src/kmer/kmc_counter/kmc_parser.hpp
@@ -13,7 +13,7 @@ namespace kmc {
 //  |min_count| -- minimum k-mer abundance (including the value passed)
 //  |max_count| -- maximum k-mer abundance (excluding the value passed)
 void read_kmers(const std::string &kmc_base_filename,
-                const std::function<void(std::string&&, uint32_t count)> &callback,
+                const std::function<void(std::string&&, uint64_t count)> &callback,
                 uint64_t min_count = 1,
                 uint64_t max_count = -1);
